transfered UIActivityViewController and ScrollView delegates
